Split graphics functions into separate file
