+++ vs/FPSci.lib.vcxproj.filters
@@ -143,6 +143,9 @@
     <ClCompile Include="..\source\ExperimentConfig.cpp">
       <Filter>Source Files\Configuration</Filter>
     </ClCompile>
+    <ClCompile Include="..\source\FPSciGraphics.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="resources.rc">
